bug-fix: Pass a copy index to obs_utitilites_mod
fixes issue #220.  Need to check create_obs is correct for other uses.
This gives the same answer as compiling make_fake_obs without debuging
options.

The subroutines create_3d_sphere_obs and create_3d_cartesian_obs
assume that the observation has a single qc value and a single data
value.  This is not true for make_fake_obs and so gave the array mismatch
described in issue #220.